Configure vimrc to support host-specific settings.
This should resolve some of my issues with different configurations for
each computer. To accomplish this, vimrc looks for a locally-created
.hostname.vimrc file before setting defaults.

Using a file on the computer rather than checking the
output of hostname() has two other advantages: it respects DRY, and
it does not expose the names of the computers I use in my vimrc (since
this is a public repository).