Create a gvimrc for graphical usage.
Apparently some installations of vim only enable clientserver in gvim.
As such, I will occasionally be forced to use a terminal application
inside of a needless GUI application with lots of overhead... such is
the state of the world. To make the transition bearable I created a
gvimrc for customizing the GUI to my liking.