Format MySessions.vim for readability.
This is in preparation for future modifications.

Now that I'm considering a tmux-based workflow, I need to be able to
accommodate multiple sessions "touching" the same directory. As such, I
will modify MySessions.vim to provide functionality for autogeneration
and selection from different sessions in the same directory... if
possible.