Purge eclim from vim
This branch exists to identify what might be wrong with my setup and
maybe fix it. I have consistently found that eclim is the most
disruptive plugin in my lineup, so removing it was a first priority.
Despite how much easier Eclim makes it to work with Java projects, it can
be a hassle to deal with.