Add a mapping to replace EOL spaces
This mapping overwrites the previous mapping for <leader>ss. It was
originally mapped to 'set spell!', but since vim-unimpaired includes an
identical, idiomatic mapping for the same functionality I didn't need it
anymore.

I frequently have problems with lines that end with spaces, so I changed
it to a mapping that deletes all line-ending spaces.