Add extra TeX compile command for LaTeX
I modified the TexC command to show errors only when the command fails,
and display the compiled pdf if it succeeds. A side effect of the
current implementation is that it still open up an extra window,
requiring a few extra keystrokes to close it; I will have to fix this in
the future.