Change Java syntax highlighting for final keyword
This commit shades the "final" keyword in any Java file the same color
as comments. This is mostly because the word "final" is proliferating in
my code the more I learn about Java and program design, and quite
frankly I don't want to see that junk at the same level of emphasis as
"abstract" or "protected". In the colorschemes I use, the color of
comments de-emphasizes text, so I figured it would be a good idea to
shade "final" like a comment (since it is essentially a documentation
elmeent anyways).